Add Dynamically sized registers (#6)
This PR adds an optimization to use dynamic register sizes. For example, if R=17, then we can use an `int` to represent a register instead of a `long` (since we'd need 6 bits for HLL and 17 for Minhash). This can have a meaningful impact when maintaining many sketches with a large number of registers in total. Since we are pre-1.0, this PR breaks the current serialization format.