IMP optimize serde and union operations
Serialization of sparse objects creates a lot of zeros, which becomes
costly in some applications when serialization is used as an
intermediate step while building up unions. The proposed scheme is
backwards-compatible with the current implementation, but also makes use
of negative values (which shouldn't be in the registers) to encode
run-lengths of zero values.

Also introduced here is an optimization for union operations, which
previously applied logic to the first sketch twice. To that end, a fix
to `deepCopy()` was also needed for `BetaMinHash`, whose `deepCopy`
never actually copied values.

Motivation:

We want to use this library in both spark and elasticsearch (as a
custom plugin); spark to create binary serialized forms and
elasticsearch to search/aggregate over them in realtime. Our Spark UDAF
necessarily has to create many sparse objects before merging them, and
this is made much more efficient in terms of runtime by finding a better
encoding scheme for such sparse objects. There is no penalty when
objects are "dense" (in terms of serialized size). Impact on
deserialization time is minimal; each value is checked for negativity.
Serialization now requires two passes: one to determine output size and
one to do the serialization.